finish writing animation logic
Got to say, I'm pleased as can be with how this came out.
